@@ -20,7 +20,7 @@ dynamic = [
 ]
 description = "A research project for emerging optimizers other than AdamW"
 license = {file = "LICENSE"}
-requires-python = ">=3.10"
+requires-python = ">=3.12"
 authors = [{ name = "NVIDIA", email = "nemo-toolkit@nvidia.com" }]
 maintainers = [{ name = "NVIDIA", email = "nemo-toolkit@nvidia.com" }]
 keywords = [
@@ -57,7 +57,6 @@ classifiers = [
 dependencies = [
     "torch",
     "absl-py",
-    "typing-extensions",
 ]
 
 [build-system]
@@ -106,7 +105,7 @@ Homepage = "https://github.com/NVIDIA-NeMo/Emerging-Optimizers"
 [tool.ruff]
 # Match black's line length
 line-length = 119
-target-version = "py310"
+target-version = "py312"
 
 # Exclude common directories that shouldn't be linted
 exclude = [
